drm/rockchip: dsi: add ability to work as a phy instead of full dsi
The dsi controller on rk3288/rk3399 includes the dphy and its controls.
On the second dsi-controller the dphy is dual-role though, output of the
dsi1 controller or input of the isp1 controller for csi.

So add the ability to also work as a phy-driver when the dphy is used
for csi functions.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>